requirements.txt
Flask==3.1.1
flask_restx==1.3.0
Flask-CORS==6.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.1.0
mongoengine==0.29.1
marshmallow==4.0.0
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==3.1.3
pillow==11.2.1
#temeletry
prometheus-client
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.41b0

# Development dependencies
pytest==8.4.0
pytest-flask==1.3.0
pytest-cov==6.1.1
python-decouple==3.8
